Back_end/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
qrcode==7.4.2
Pillow==10.4.0
python-multipart==0.0.9
cos-python-sdk-v5==1.9.42
